Import lastest HAL from STM32Duino

+++ system/Drivers/STM32F1xx_HAL_Driver/Inc/stm32f1xx_ll_spi.h
@@ -1011,7 +1011,7 @@ __STATIC_INLINE uint32_t LL_SPI_IsEnabledDMAReq_TX(SPI_TypeDef *SPIx)
   */
 __STATIC_INLINE uint32_t LL_SPI_DMA_GetRegAddr(SPI_TypeDef *SPIx)
 {
-  return (uintptr_t) &(SPIx->DR);
+  return (uint32_t) &(SPIx->DR);
 }
 
 /**
